Move migration code from serializer.c to its own file
